# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "This is an API documentation for Posts backend",
        "title": "Swagger Posts API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "kasimovfirdavs@gmail.com"
        },
        "version": "1.0"
    },
    "host": "localhost:8080/",
    "basePath": "/",
    "paths": {
        "/health": {
            "get": {
                "description": "Get health of server",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Health-Server"
                ],
                "summary": "Get health of server",
                "operationId": "get-health",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/ports": {
            "get": {
                "description": "It will return all the available ports from the DB. We will use API caching for this purpose\nso we don't have to get all data over again from DB, which is useful in real world applications\nwhere we are connected to the real database such as PostgresSQL it saves a lot of latency.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Ports"
                ],
                "summary": "It will return all the available ports from the DB",
                "operationId": "list-ports",
                "responses": {
                    "201": {
                        "description": "Created"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            },
            "post": {
                "description": "It will create ports and save into the DB by default ports.json file will be used",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Ports"
                ],
                "summary": "It will create ports and save into the DB by default ports.json file will be used",
                "operationId": "save-ports",
                "responses": {
                    "201": {
                        "description": "Created"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/ports/from-file": {
            "post": {
                "description": "You are able to provide json file the service will parse and save into the DB",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Ports"
                ],
                "summary": "You are able to provide json file the service will parse and save into the DB",
                "operationId": "save-ports-from-file",
                "parameters": [
                    {
                        "type": "file",
                        "description": "File",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        }
    },
    "securityDefinitions": {
        "Bearer": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}